requirements.txt
fastapi==0.115.5
passlib==1.7.4
pydantic==2.10.2
PyJWT==2.10.1
sqlmodel==0.0.22
uvicorn==0.32.1
bcrypt
psycopg
psycopg-binary
httpx
requests
